+++ lib/tests/parsing.rs
@@ -358,17 +358,14 @@ mod abbreviations {
                 expected_item!(0, Long, "foo"),
             ]
         );
-        let opts = gong_option_set!(
-            [
-                // Multiple options that “foo” will match as an abbreviation for before getting to
-                // the exact match.
-                gong_longopt!("fooo"),
-                gong_longopt!("foooo"),
-                gong_longopt!("fooooo"),
-                gong_longopt!("foo"),    // Exact match for input `--foo`
-            ],
-            []
-        );
+        let opts = gong_option_set!(@long [
+            // Multiple options that “foo” will match as an abbreviation for before getting to the
+            // exact match.
+            gong_longopt!("fooo"),
+            gong_longopt!("foooo"),
+            gong_longopt!("fooooo"),
+            gong_longopt!("foo"),    // Exact match for input `--foo`
+        ]);
         let parser = Parser::new(&opts, None);
         check_result(&Actual(parser.parse(&args)), &expected);
     }
@@ -930,11 +927,9 @@ mod commands {
     /// Command names simply should not be prefixed like this.
     #[test]
     fn name_like_option() {
-        let opts = gong_option_set!(
-            [
-                gong_longopt!("foo"),
-            ], []
-        );
+        let opts = gong_option_set!(@long [
+            gong_longopt!("foo"),
+        ]);
         let cmds = gong_command_set!([
             gong_command!("--foo"),
             gong_command!("--bar"),